rename all the files based on the standards
