+++ controllers/UsersController.js
@@ -1,6 +1,6 @@
 const bcrypt = require("bcryptjs");
 const jwt = require("jsonwebtoken");
-const User = require("../models/userModel");
+const User = require("../models/User");
 
 const secretKey = process.env.SECRET_KEY;
 const users = [
