+++ routes/RoutesRoutes.js
@@ -1,6 +1,6 @@
 const express = require("express");
 const router = express.Router();
-const routeController = require("../controllers/routeController");
+const routeController = require("../controllers/RoutesController");
 
 router.get("/", routeController.getAll);
 router.get("/:id", routeController.getById);
