+++ routes/TrainsRoutes.js
@@ -1,6 +1,6 @@
 const express = require("express");
 const router = express.Router();
-const trainController = require("../controllers/trainController");
+const trainController = require("../controllers/TrainsController");
 
 router.get("/", trainController.getAll);
 router.get("/:id", trainController.getById);
